+++ CHANGES.md
@@ -10,6 +10,9 @@ The released versions correspond to PyPI releases.
   (see [#1300](https://github.com/pytest-dev/pyfakefs/issues/1300))
 * fake file `seek` method did not return the location in the file
   (see [#1304](https://github.com/pytest-dev/pyfakefs/issues/1304))
+* make sure case sensitivity is correctly set for fake posix paths
+  in `hash()`, `Path.match` and `Path.full_match`
+  (see [#1308](https://github.com/pytest-dev/pyfakefs/issues/1308))
 
 ## [Version 6.1.6](https://pypi.python.org/pypi/pyfakefs/6.1.6) (2026-03-18)
 Follow-up bugfix release for 6.1.5.